chore: adds ability to switch between game states
Instead of all the code for the game going into one file, we now can
split the code between different files via states or scenes essentially.

So now the game starts in the "Main Menu" state and the player can
use the up and down arrows to select an option (no text yet lol).

Only the Practice Mode and Quit options are implemented currently.

part of HEX-33

Signed-off-by: Jack Kinsey <jack.w.kinsey@gmail.com>